+++ DTCD-WorkspaceSystem/rollup.config.js
@@ -3,6 +3,7 @@ import commonjs from '@rollup/plugin-commonjs';
 import resolve from '@rollup/plugin-node-resolve';
 import babel from '@rollup/plugin-babel';
 import json from '@rollup/plugin-json';
+import html from 'rollup-plugin-html';
 
 const watch = Boolean(process.env.ROLLUP_WATCH);
 
@@ -15,6 +16,7 @@ const fileDest = watch
 const plugins = [
   resolve(),
   commonjs(),
+  html({ include: '**/*.html' }),
   babel({ babelHelpers: 'bundled' }),
   styles({ mode: 'inject' }),
   json(),